+++ src/projects/repeat_resolution/CMakeLists.txt
@@ -1,4 +1,4 @@
 project(repeat_resolution CXX)
 
-add_library(repeat_resolution STATIC paths.cpp multiplex_dbg_topology.cpp)
-target_link_libraries(repeat_resolution graphlite)
+add_library(repeat_resolution STATIC paths.cpp mdbg_topology.cpp mdbg_inc.cpp mdbg_vertex_processor.cpp mdbg_vertex_processor.hpp mdbg.cpp)
+target_link_libraries(repeat_resolution graphlite lja_dbg)
